pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "matplotlib>=3.10.6",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
